% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCV.R
\name{calcCV}
\alias{calcCV}
\title{calcCV}
\usage{
calcCV(avgCtObj, cutoff = 1)
}
\arguments{
\item{avgCtObj}{A data.frame generated from the \code{\link{getAvgCT}} function.}

\item{cutoff}{The cutoff value for the coefficient of variation for triggering a warning message. Default 1.0, or 1\%.}
}
\value{
A data.frame.
}
\description{
Calculates the coefficient of variation for each gene for each sample and triggers a message if any samples
are above the desired threshold.
}
\details{
Technical replicate error is not carried forward when calculating error for biological replicates. The coefficient of variation
is intended to be used a filter for checking the quality of technical replication before proceeding.
}
\examples{

}
