% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcStab.R
\name{calcStab}
\alias{calcStab}
\title{calcStab}
\usage{
calcStab(avgCTobj, hkgs)
}
\arguments{
\item{avgCTobj}{A data.frame generated from the \code{\link{getAvgCT}} function.}

\item{hkgs}{A character vector containing the potential housekeeping genes.}
}
\value{
A data.frame containing the tested housekeeping genes and their scores,
sorted from lowest to highest (i.e. most stable to least stable).
}
\description{
Finds the most stable housekeeping gene across conditions using the DeltaCt approach from Silver et al. (2006).
}
\examples{
\dontrun{
geneRank <- calcStab(avgCTobj, c("ACTB", "GAPDH"))
}

}
\references{
\href{https://www.ncbi.nlm.nih.gov/pubmed/17026756}{Silver et al. 2006. Selection of housekeeping genes for gene expression studies in human reticulocytes using real-time PCR. BMC Mol Biol}
}
