% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_chip_qpcr.R
\name{do_chip_qpcr}
\alias{do_chip_qpcr}
\title{do_chip_qpcr}
\usage{
do_chip_qpcr(sampleObj, inputName = NULL, mockName = NULL,
  inputPercent = NULL, method, bioReps = FALSE)
}
\arguments{
\item{sampleObj}{A qPCRBatch object generated from the \code{\link{readSampleSheet}} function.}

\item{inputName}{A character vector of the common name given to all input samples}

\item{mockName}{A character vector of the common name given to all mock (e.g. IgG control) samples}

\item{inputPercent}{The percent of starting chromatin used as input. Should be an integer in percent (e.g. 2\% input is 2).}

\item{method}{Indicates which calculation should be performed. Must be one of ("i", "input") for input,
or ("fe", "fold enrichment") for fold enrichment.}

\item{bioReps}{Boolean indicating whether or not there are biological replicates.}
}
\value{
A data.frame with dCT values, percent input/fold enrichment, and max/min percent input/fold enrichment per sample.
}
\description{
Calcuates either percent input or fold enrichment for chip qpcr data
}
\details{
This function will handle both single replicate or n > 1 replicate experiments. Providing a `mockName` for percent
input calculations will remove the mock samples from the calculations. Likewise, providing an `inputName` for fold enrichment
calculations will remove the input samples from the calculations.
}
