% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_chip_dCT.R
\name{get_chip_dCT}
\alias{get_chip_dCT}
\title{get_chip_dCT}
\usage{
get_chip_dCT(sampleObj, inputName = NULL, mockName = NULL,
  inputPercent = NULL, method)
}
\arguments{
\item{sampleObj}{A qPCRBatch object generated from the \code{\link{readSampleSheet}} function.}

\item{inputName}{A character vector of the common name given to all input samples}

\item{mockName}{A character vector of the common name given to all mock (e.g. IgG control) samples}

\item{inputPercent}{The percent of starting chromatin used as input. Should be an integer in percent (e.g. 2\% input is 2).}

\item{method}{Indicates which calculation should be performed. Must be one of ("i", "input") for input,
or ("fe", "fold enrichment") for fold enrichment.}
}
\value{
A data.frame
}
\description{
Calculates dCT values for chip qpcr data and organizes it into a format
compatible with \code{\link{signifTest}}.
}
\details{
This function only works with biological replicate data. Providing a `mockName` for percent
input calculations will remove the mock samples from the calculations. Likewise, providing an `inputName` for
fold enrichment calculations will remove the input samples from the calculations.
}
