% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dCT_bioReps.R
\name{get_dCT_bioReps}
\alias{get_dCT_bioReps}
\title{get_dCT_bioReps}
\usage{
get_dCT_bioReps(sampleObj, hkg)
}
\arguments{
\item{sampleObj}{A qPCRBatch object generated from the \code{\link{readSampleSheet}} function.}

\item{hkg}{A character vector containing the housekeeping gene to use for dCT calculations.}
}
\value{
A data.frame containing dCT values for each condition.
}
\description{
Calculates dCT values when biological replicates are present.
}
\details{
Biological replicates must have unique identifiers, else they will be considered technical replicates.
The simplest way to do this is to give samples names such as "treated_1", "treated_2", etc.
}
\examples{

}
