% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ddCT_bioReps.R
\name{get_ddCT_bioReps}
\alias{get_ddCT_bioReps}
\title{get_ddCT_bioReps}
\usage{
get_ddCT_bioReps(sampleObj, cond.1, cond.2, reps.cond.1, reps.cond.2, hkg,
  rel.exp = FALSE)
}
\arguments{
\item{sampleObj}{A qPCRBatch object generated from the \code{\link{readSampleSheet}} function.}

\item{cond.1}{A character vector containing the group for condition 1 (e.g. "untreated").}

\item{cond.2}{A character vector containing the group for condition 1 (e.g. "treated").}

\item{reps.cond.1}{The number of biological replicates in condition 1.}

\item{reps.cond.2}{The number of biological replicates in condition 2.}

\item{hkg}{A character vector containing the housekeeping gene to use for ddCT calculations.}

\item{rel.exp}{Boolean value indicating whether or not relative expression values should be calculated.
TRUE will ONLY compute expression relative to the control sample,
FALSE will ONLY compute log2 fold-change values.}
}
\value{
A data.frame containing ddCT values in log2 fold-change or relative expression format.
}
\description{
Calculates the ddCT values for all samples when biological replicates are present.
}
\examples{
\dontrun{
sampleObj <- readSampleSheet(file)
ddCTObj <- get_ddCT_bioReps(sampleObj, cond.1 = "untreated",
cond.2 = "treated", reps.cond.1 = 3, reps.cond.2 = 3,
hkg = "GAPDH", rel.exp = FALSE)
}

}
