% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ddCT_singleRep.R
\name{get_ddCT_singleRep}
\alias{get_ddCT_singleRep}
\title{get_ddCT_singleRep}
\usage{
get_ddCT_singleRep(sampleObj, hkg, control, rel.exp = FALSE)
}
\arguments{
\item{sampleObj}{A qPCRBatch object generated from the \code{\link{readSampleSheet}} function.}

\item{hkg}{A character vector containing the housekeeping gene to use for ddCT calculations.}

\item{control}{A character vector containing the name of the control condition.}

\item{rel.exp}{Boolean value indicating whether or not relative expression values should be calculated.
TRUE will compute both log2 fold-change and expression relative to the control sample,
FALSE will only compute log2 fold-change values.}
}
\value{
A data.frame containing ddCT values in log2 fold-change or relative expression format.
}
\description{
Calculates the ddCT values for all samples in situations where
there is only a single biological replicate per condition.
}
\details{
This function is primarily intended for pilot/preliminary data in which biological replicates have not
yet been produced or cases in which multiple vectors are being evaluated for efficiency (e.g. shRNA knockdowns, overexpression
constructs, etc.)
}
\examples{
\dontrun{
sampleObj <- readSampleSheet(file)
ddCTObj <- get_ddCT_singleRep(sampleObj, "GAPDH", "untreated",
rel.exp = FALSE)
}

}
