% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAvgCT.R
\name{getAvgCT}
\alias{getAvgCT}
\title{getAvgCT}
\usage{
getAvgCT(sampleObj)
}
\arguments{
\item{sampleObj}{A qPCRBatch object.}
}
\value{
A data.frame.
}
\description{
Calculates the average CT value and standard deviation from technical replicates.
}
\examples{

}
