% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRawCT.R
\name{getRawCT}
\alias{getRawCT}
\title{getRawCT}
\usage{
getRawCT(sampleObj)
}
\arguments{
\item{sampleObj}{A qPCRBatch object.}
}
\value{
A data.frame.
}
\description{
Generates a data.frame of raw CT values.
}
\examples{

}
