% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NAtoNumeric.R
\name{NAtoNumeric}
\alias{NAtoNumeric}
\title{NAtoNumeric}
\usage{
NAtoNumeric(sampleObj, value = 35)
}
\arguments{
\item{sampleObj}{The qPCRBatch object from \code{\link{readSampleSheet}} to be converted.}

\item{value}{Value to replace NA with. Default 35.}
}
\value{
A qPCRBatch object with NAs replaced with \code{value}.
}
\description{
Takes a qPCRBatch object as input and converts all Cq/Ct NAs to desired values.
}
\examples{

}
