% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_chip_qpcr.R
\name{plot_chip_qpcr}
\alias{plot_chip_qpcr}
\title{plot_chip_qpcr}
\usage{
plot_chip_qpcr(chip_qpcr_df, palette = "Set1", xlab = "", ylab = "",
  title = "", legend.title = "", theme_classic = TRUE,
  method = NULL, hideInput = FALSE, inputName, hideMock = FALSE,
  mockName)
}
\arguments{
\item{chip_qpcr_df}{The data.frame generated by the \code{\link{do_chip_qpcr}} function}

\item{palette}{Any color palette from RColorBrewer}

\item{xlab}{x axis label}

\item{ylab}{y axis label}

\item{title}{Main title for plot}

\item{legend.title}{Title for plot legend}

\item{theme_classic}{Boolean value. TRUE will yield a plot with the 'classic' background. FALSE will yield a plot
with the default ggplot2 theme.}

\item{method}{Indicates which calculation should be performed. Must be one of ("i", "input") for input,
or ("fe", "fold enrichment") for fold enrichment.}

\item{hideInput}{Boolean value indicating whether or not the input samples themselves should be removed from the plot}

\item{inputName}{A character vector of the common name given to all input samples}

\item{hideMock}{Boolean value indicating whether or not the mock samples themselves should be removed from the plot}

\item{mockName}{A character vector of the common name given to all mock samples}
}
\value{
A ggplot2 barplot
}
\description{
plot_chip_qpcr
}
\details{
The `chip_qpcr_df` object passed to the function must be in the same format generated by the
\code{\link{do_chip_qpcr}} function.
}
\examples{

}
