% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dCT_singleRep.R
\name{plot_dCT_singleRep}
\alias{plot_dCT_singleRep}
\title{plot_dCT_singleRep}
\usage{
plot_dCT_singleRep(dCTobj, palette = "Set1", xlab = "", ylab = "",
  title = "", legend.title = "", theme_classic = TRUE)
}
\arguments{
\item{dCTobj}{A data.frame generated from the \code{\link{get_dCT_singleRep}} function.}

\item{palette}{Any color palette from RColorBrewer.}

\item{xlab}{x axis label.}

\item{ylab}{y axis label.}

\item{title}{Main title for the plot.}

\item{legend.title}{Title for the plot legend.}

\item{theme_classic}{Boolean value. TRUE will yield a plot with the 'classic' background. FALSE will yield a plot
with the default ggplot2 theme.}
}
\value{
A ggplot2 barplot of dCT values.
}
\description{
Generates a barplot of the dCT values for each sample with error bars.
}
\examples{

}
