% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ddCT_bioReps.R
\name{plot_ddCT_bioReps}
\alias{plot_ddCT_bioReps}
\title{plot_ddCT_bioReps}
\usage{
plot_ddCT_bioReps(ddCTobj, palette = "Set1", xlab = "Gene",
  ylab = "", title = "", legend.title = "", hideHKG = FALSE,
  HKGname, theme_classic = TRUE, rel.exp = FALSE)
}
\arguments{
\item{ddCTobj}{A data.frame generated from the \code{\link{get_ddCT_bioReps}} function.}

\item{palette}{Any color palette from RColorBrewer.}

\item{xlab}{x axis label.}

\item{ylab}{y axis label.}

\item{title}{Main title for the plot.}

\item{legend.title}{Title for the plot legend.}

\item{hideHKG}{Boolean value indicating whether or not the housekeeping gene should be removed from the plot}

\item{HKGname}{A character vector of the housekeeping gene name. Required if `hideHKG` is TRUE.}

\item{theme_classic}{Boolean value. TRUE will yield a plot with the 'classic' background. FALSE will yield a plot
with the default ggplot2 theme.}

\item{rel.exp}{Boolean value indicating the data type contained in the ddCTobj.
TRUE indicates the data is on the relative expression scale and will generate
a plot of relative expression values.
FALSE indicates the data is on the log2 fold-change scale will generate
a plot of log2 fold-change values.}
}
\value{
A ggplot2 barplot of ddCT values.
}
\description{
Generates a barplot of the ddCT values for each condition with error bars.
}
\details{
This function requires the data contained in the ddCTobj to match the setting chosen
for the `rel.exp` argument. Mismatches (i.e. having a ddCTobj with relative expression
and requesting a plot of log2 fold-change) will result in errors.
}
\examples{

}
