% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ddCT_singleRep.R
\name{plot_ddCT_singleRep}
\alias{plot_ddCT_singleRep}
\title{plot_ddCT_singlerep}
\usage{
plot_ddCT_singleRep(ddCTobj, palette = "Set1", xlab = "", ylab = "",
  title = "", legend.title = "", theme_classic = TRUE,
  rel.exp = FALSE)
}
\arguments{
\item{ddCTobj}{A data.frame generated from the \code{\link{get_ddCT_singleRep}} function.}

\item{palette}{Any color palette from RColorBrewer.}

\item{xlab}{x axis label.}

\item{ylab}{y axis label.}

\item{title}{Main title for the plot.}

\item{legend.title}{Title for the plot legend.}

\item{theme_classic}{Boolean value. TRUE will yield a plot with the 'classic' background. FALSE will yield a plot
with the default ggplot2 theme.}

\item{rel.exp}{Boolean value indicating whether or not relative expression values should be used for plot. TRUE will generate
a plot of relative expression values, FALSE will generate a plot of log2 fold-change values.}
}
\value{
A ggplot2 barplot of ddCT values.
}
\description{
Generates a barplot of the ddCT values for each sample with error bars.
}
\examples{

}
