% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAvgCT.R
\name{plotAvgCT}
\alias{plotAvgCT}
\title{plotAvgCT}
\usage{
plotAvgCT(avgCtObj, palette = "Set1", xlab = "", ylab = "",
  title = "", legend.title = "", theme_classic = TRUE)
}
\arguments{
\item{avgCtObj}{A data.frame generated from the \code{\link{getAvgCT}} function.}

\item{palette}{Any color palette from RColorBrewer.}

\item{xlab}{x axis label.}

\item{ylab}{y axis label.}

\item{title}{Main title for the plot.}

\item{legend.title}{Title for the plot legend.}

\item{theme_classic}{Boolean value. TRUE will yield a plot with the 'classic' background. FALSE will yield a plot
with the default ggplot2 theme.}
}
\value{
A ggplot2 barplot of average CT values.
}
\description{
Generates a barplot of the average CT values for each sample with error bars.
}
\examples{

}
