% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSampleSheet.R
\name{readSampleSheet}
\alias{readSampleSheet}
\title{readSampleSheet}
\usage{
readSampleSheet(file)
}
\arguments{
\item{file}{}
}
\value{
A qPCRBatch object.
}
\description{
Reads in a tab-delimited file of qPCR data using \code{\link{read.qPCR}}.
Requires 5 columns: Well, Plate, Sample, Detector, Cq.
}
\details{
Biological replicates MUST be given unique identifiers, else they will be treated as technical replicates. For example,
if an experiment has 2 biological replicates for a wildtype sample, they should be entered as "WT_1" and "WT_2". Entering them
both as "WT" will result in them being treated as technical replicates of the same sample. Multiple technial replicates of biological
replicates should entered with the same names (e.g. "WT_1", "WT_1", "WT_2", "WT_2" would be 2 technical replicates of biological
replicate 1 and 2 technical replicates of biological replicate 2).
}
\examples{

}
