% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signifTest.R
\name{signifTest}
\alias{signifTest}
\title{signifTest}
\usage{
signifTest(dCTObj, gene.name, var.equal = TRUE)
}
\arguments{
\item{dCTObj}{A data.frame generated from the \code{\link{get_dCT_singleRep}}
or \code{\link{get_dCT_bioReps}} function.}

\item{gene.name}{Name of the gene for which to compute the t-test.}

\item{var.equal}{Boolean indicating whether or not to treat the variance as equal.}
}
\value{
A list containing the components indicated in \code{\link{t.test}}. Also prints a message
reporting the calculated p.value.
}
\description{
Computes a p.value using a two-sample t-test with H0 = "two.sided" for the indicated gene.
See \code{\link{t.test}} for more information.
}
\examples{

}
